Convert template switch to use async_track_template_result
